support multi-line strings
Multi-line strings break unfolding rules of RFC 5545, as new lines are
expected to be new content lines, or start with whitespace (which would
indicate that it is part of a string from the previous content line).

Google Calendar's ICS output just replaces newlines with `\n` in strings
and then folds as normal - so this copies that.

Signed-off-by: David Wood <david@davidtw.co>