Fix path handling for included calendar files
When this tool is run from a different directory than where the
calendar files are located, it fails if one calendar file includes
another.

This happens because we're not extracting the base path from the input
file and prefixing the included files with that path.

Let's do that so that this tool can be run from an arbitrary
directory.