Handle `calendar` repo containing `xtask` in CI
The CI iterates over and calls `toml-to-ical` on all `.toml` files in
the `calendar` repo.  However, we'd like to add an `xtask` directory
to the `calendar` repo, and that directory will contain non-calendar
TOML files.  We may also want a top-level `Cargo.toml` file.

To support this, let's skip any files named `Cargo.toml` (a calendar
for the cargo team would be in lowercase) and a top-level directory
named `xtask`.